+++ packages/web3-eth-personal/test/unit/eth_personal.test.ts
@@ -15,7 +15,7 @@ You should have received a copy of the GNU Lesser General Public License
 along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-import { ETH_DATA_FORMAT } from 'web3-utils';
+import { ETH_DATA_FORMAT } from 'web3-types';
 import * as utils from 'web3-utils';
 import * as eth from 'web3-eth';
 import * as validator from 'web3-validator';