+++ packages/web3-eth/src/index.ts
@@ -22,7 +22,7 @@ along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
  * ```ts
  * import { Web3Context } from 'web3-core';
  * import { BlockTags } from 'web3-types';
- * import { DEFAULT_RETURN_FORMAT } from 'web3-utils';
+ * import { DEFAULT_RETURN_FORMAT } from 'web3-types';
  * import { getBalance} from 'web3-eth';
  *
  * getBalance(