+++ packages/web3-eth/src/utils/detect_transaction_type.ts
@@ -15,9 +15,9 @@ You should have received a copy of the GNU Lesser General Public License
 along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-import { ETH_DATA_FORMAT, format, toHex } from 'web3-utils';
+import { format, toHex } from 'web3-utils';
 import { TransactionTypeParser, Web3Context } from 'web3-core';
-import { EthExecutionAPI, Transaction } from 'web3-types';
+import { EthExecutionAPI, Transaction, ETH_DATA_FORMAT } from 'web3-types';
 import { isNullish } from 'web3-validator';
 import { InternalTransaction } from '../types';