+++ packages/web3-eth/src/utils/watch_transaction_by_subscription.ts
@@ -15,8 +15,9 @@ You should have received a copy of the GNU Lesser General Public License
 along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
 import { Bytes, Numbers, BlockHeaderOutput, TransactionReceipt } from 'web3-types';
-import { DataFormat, format } from 'web3-utils';
+import { format } from 'web3-utils';
 
+import { DataFormat } from 'web3-types';
 import { NewHeadsSubscription } from '../web3_subscriptions';
 import { transactionReceiptSchema } from '../schemas';
 import { WaitProps, watchTransactionByPolling } from './watch_transaction_by_pooling';