+++ packages/web3-eth/test/fixtures/detect_transaction_type.ts
@@ -15,8 +15,7 @@ You should have received a copy of the GNU Lesser General Public License
 along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-import { ETH_DATA_FORMAT, FormatType } from 'web3-utils';
-import { Transaction } from 'web3-types';
+import { ETH_DATA_FORMAT, FormatType, Transaction } from 'web3-types';
 
 export const transactionType0x0: FormatType<Transaction, typeof ETH_DATA_FORMAT>[] = [
 	{