+++ packages/web3-eth/test/fixtures/validate_transaction_for_signing.ts
@@ -15,8 +15,7 @@ You should have received a copy of the GNU Lesser General Public License
 along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-import { Transaction } from 'web3-types';
-import { FMT_BYTES, FMT_NUMBER, FormatType } from 'web3-utils';
+import { Transaction, FMT_BYTES, FMT_NUMBER, FormatType } from 'web3-types';
 import {
 	ChainIdMismatchError,
 	CommonOrChainAndHardforkError,