+++ packages/web3-eth/test/unit/rpc_method_wrappers/sign_transaction.test.ts
@@ -15,9 +15,13 @@ You should have received a copy of the GNU Lesser General Public License
 along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
 import { Web3Context } from 'web3-core';
-import { DEFAULT_RETURN_FORMAT, ETH_DATA_FORMAT } from 'web3-utils';
+import {
+	DEFAULT_RETURN_FORMAT,
+	ETH_DATA_FORMAT,
+	SignedTransactionInfoAPI,
+	Web3EthExecutionAPI,
+} from 'web3-types';
 import { isString } from 'web3-validator';
-import { SignedTransactionInfoAPI, Web3EthExecutionAPI } from 'web3-types';
 import { ethRpcMethods } from 'web3-rpc-methods';
 
 import { signTransaction } from '../../../src/rpc_method_wrappers';