+++ packages/web3-eth/test/unit/utils/watch_transaction_for_confirmations.test.ts
@@ -15,8 +15,8 @@ You should have received a copy of the GNU Lesser General Public License
 along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
 import { Web3Context, Web3PromiEvent } from 'web3-core';
-import { DEFAULT_RETURN_FORMAT, format } from 'web3-utils';
-import { TransactionReceipt, Web3EthExecutionAPI } from 'web3-types';
+import { format } from 'web3-utils';
+import { DEFAULT_RETURN_FORMAT, TransactionReceipt, Web3EthExecutionAPI } from 'web3-types';
 import { ethRpcMethods } from 'web3-rpc-methods';
 import {
 	TransactionMissingReceiptOrBlockHashError,