+++ packages/web3-utils/test/fixtures/formatter.ts
@@ -14,7 +14,8 @@ GNU Lesser General Public License for more details.
 You should have received a copy of the GNU Lesser General Public License
 along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
-import { FMT_NUMBER, FMT_BYTES } from '../../src/formatter';
+
+import { FMT_NUMBER, FMT_BYTES } from 'web3-types';
 
 export const isDataFormatValid: [any, boolean][] = [
 	[{ number: 'number', bytes: 'number' }, true],