+++ packages/web3-eth/CHANGELOG.md
@@ -118,6 +118,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   `signTransaction` will now return `gas` instead of `gasLimit` for returned transaction object regardless of what property name the provider uses (#5915)
 -   `formatTransaction` will now replace `data` transaction property with `input` (#5915)
 -   `isTransactionCall` will now check if `value.input` `isHexStrict` if provided (#5915)
+-   The functions `defaultTransactionBuilder` and `transactionBuilder` are now utilizing the type `Transaction` for the transaction object. (#5993)
 
 ### Added