+++ packages/web3-utils/CHANGELOG.md
@@ -96,6 +96,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Added source files (#5956)
 -   Added hybrid build (ESM and CJS) of library (#5904)
 
+### Changed
+
+-   The types `FMT_NUMBER`, `NumberTypes`, `FMT_BYTES`, `ByteTypes`, `DataFormat`, `DEFAULT_RETURN_FORMAT`, `ETH_DATA_FORMAT` and `FormatType` moved to `web3-types`. (#5993)
+
 ### Removed
 
 -   Removed dependencies @ethereumjs/tx, @ethereumjs/common (#5963)